requirements.txt
beautifulsoup4
mysql-connector-python
requests
googletrans==4.0.0rc1
pandas
plotly
